+++ server/src/solver/solver/graph.c
@@ -3,7 +3,7 @@
 #include <time.h>
 #include <string.h>
 
-int initialiser_graphe(Graphe* graphe, int nombre_sommets, int nombre_aretes) {
+int initialiser_graphe(Graphe* graphe, long nombre_sommets, long nombre_aretes) {
     graphe->nombre_sommets = nombre_sommets;
     graphe->nombre_aretes = nombre_aretes;
     graphe->sommets = malloc(sizeof(Sommet) * (nombre_sommets + 1)); // id 0 non utilisé
@@ -15,14 +15,14 @@ int initialiser_graphe(Graphe* graphe, int nombre_sommets, int nombre_aretes) {
         return 1;
     }
 
-    for (int i = 0; i <= nombre_sommets; i++) {
+    for (long i = 0; i <= nombre_sommets; i++) {
         graphe->sommets[i].id_article = -1;
 		graphe->sommets[i].id_graphe = -1;
         graphe->sommets[i].debut_in_array = -1;
         graphe->sommets[i].nombre_aretes = 0;
     }
 
-    for (int i = 0; i < nombre_aretes; i++) {
+    for (long i = 0; i < nombre_aretes; i++) {
         graphe->aretes[i].id_article_depart = -1;
         graphe->aretes[i].id_article_arrival = -1;
     }
@@ -59,8 +59,8 @@ int lire_fichier_et_creer_graphe(const char* chemin_fichier, Graphe* graphe) {
         return 1;
     }
 
-    int nombre_sommets, nombre_aretes;
-    if (sscanf(ligne, "%d %d", &nombre_sommets, &nombre_aretes) != 2) {
+    long nombre_sommets, nombre_aretes;
+    if (sscanf(ligne, "%ld %ld", &nombre_sommets, &nombre_aretes) != 2) {
         fprintf(stderr, "Error: file not correctly formatted.\n");
         fclose(fichier);
         return 1;
@@ -70,8 +70,8 @@ int lire_fichier_et_creer_graphe(const char* chemin_fichier, Graphe* graphe) {
         return 1;
 
     int lecture_aretes = 0;
-    int idx_sommet = 0;
-    int idx_arete = 0;
+    long idx_sommet = 0;
+    long idx_arete = 0;
 
     while (fgets(ligne, sizeof(ligne), fichier)) {
         if (ligne[0] == '-') {
@@ -80,8 +80,8 @@ int lire_fichier_et_creer_graphe(const char* chemin_fichier, Graphe* graphe) {
         }
 
         if (!lecture_aretes) {
-            int id_article, id_graphe, debut, nb;
-            if (sscanf(ligne, "%d %d %d %d", &id_article, &id_graphe, &debut, &nb) == 4) {
+            long id_article, id_graphe, debut, nb;
+            if (sscanf(ligne, "%ld %ld %ld %ld", &id_article, &id_graphe, &debut, &nb) == 4) {
                 Sommet* s = &graphe->sommets[idx_sommet];
                 s->id_article     = id_article;
                 s->id_graphe      = id_graphe;
@@ -97,8 +97,8 @@ int lire_fichier_et_creer_graphe(const char* chemin_fichier, Graphe* graphe) {
                 idx_sommet++;
             }
         } else {
-            int from, to;
-            if (sscanf(ligne, "%d %d", &from, &to) == 2) {
+            long from, to;
+            if (sscanf(ligne, "%ld %ld", &from, &to) == 2) {
                 graphe->aretes[idx_arete].id_article_depart = from;
                 graphe->aretes[idx_arete].id_article_arrival= to;
                 idx_arete++;
@@ -111,24 +111,24 @@ int lire_fichier_et_creer_graphe(const char* chemin_fichier, Graphe* graphe) {
     return 0;
 }
 
-int find_id_graphe(const Graphe* graphe, int id_article) {
+int find_id_graphe(const Graphe* graphe, long id_article) {
     MapEntry* e;
     HASH_FIND_INT(graphe->map_id, &id_article, e);
     return e ? e->id_graphe : -1;
 }
 
 void afficher_graphe(const Graphe* graphe) {
-    printf("Graphe : %d sommets, %d arêtes\n",
+    printf("Graphe : %ld sommets, %ld arêtes\n",
            graphe->nombre_sommets, graphe->nombre_aretes);
-    for (int i = 0; i < graphe->nombre_sommets; i++) {
+    for (long i = 0; i < graphe->nombre_sommets; i++) {
         Sommet* s = &graphe->sommets[i];
-        printf("Sommet id_article=%d, id_graphe=%d, %d arêtes, debut=%d\n",
+        printf("Sommet id_article=%ld, id_graphe=%ld, %ld arêtes, debut=%ld\n",
                s->id_article, s->id_graphe,
                s->nombre_aretes, s->debut_in_array);
     }
-    for (int i = 0; i < graphe->nombre_aretes; i++) {
+    for (long i = 0; i < graphe->nombre_aretes; i++) {
         Arete* a = &graphe->aretes[i];
-        printf("Arete %d: %d -> %d\n",
+        printf("Arete %ld: %ld -> %ld\n",
                i, a->id_article_depart, a->id_article_arrival);
     }
 }
@@ -141,20 +141,20 @@ void save_graphe_binaire(const Graphe* g, const char* path) {
     }
 
     // Header info
-    fwrite(&g->nombre_sommets, sizeof(int), 1, f);
-    fwrite(&g->nombre_aretes,  sizeof(int), 1, f);
+    fwrite(&g->nombre_sommets, sizeof(long), 1, f);
+    fwrite(&g->nombre_aretes,  sizeof(long), 1, f);
 
     // Graph info
     fwrite(g->sommets, sizeof(Sommet), g->nombre_sommets+1, f);
     fwrite(g->aretes, sizeof(Arete), g->nombre_aretes, f);
 
     // Hash table info
-    unsigned int map_size = HASH_COUNT(g->map_id);
-    fwrite(&map_size, sizeof(unsigned int), 1, f);
+    long map_size = HASH_COUNT(g->map_id);
+    fwrite(&map_size, sizeof(long), 1, f);
     MapEntry* entry;
     for (entry = g->map_id; entry; entry = entry->hh.next) {
-        fwrite(&entry->id_article, sizeof(int), 1, f);
-        fwrite(&entry->id_graphe,  sizeof(int), 1, f);
+        fwrite(&entry->id_article, sizeof(long), 1, f);
+        fwrite(&entry->id_graphe,  sizeof(long), 1, f);
     }
 
     fclose(f);
@@ -168,8 +168,8 @@ int load_graphe_binaire(const char* path, Graphe* g) {
         return 1;
     }
 
-    int nbS, nbA;
-    if (fread(&nbS, sizeof(int), 1, f)!=1 || fread(&nbA, sizeof(int), 1, f)!=1) {
+    long nbS, nbA;
+    if (fread(&nbS, sizeof(long), 1, f)!=1 || fread(&nbA, sizeof(long), 1, f)!=1) {
         fprintf(stderr, "Error getting header info\n"); fclose(f); return 1;
     }
     initialiser_graphe(g, nbS, nbA);
@@ -177,12 +177,12 @@ int load_graphe_binaire(const char* path, Graphe* g) {
     fread(g->sommets, sizeof(Sommet), nbS+1, f);
     fread(g->aretes,  sizeof(Arete),  nbA,   f);
 
-    unsigned int map_size;
-    fread(&map_size, sizeof(unsigned int), 1, f);
-    for (unsigned int i = 0; i < map_size; i++) {
+    long map_size;
+    fread(&map_size, sizeof(long), 1, f);
+    for (long i = 0; i < map_size; i++) {
         MapEntry* e = malloc(sizeof(MapEntry));
-        fread(&e->id_article, sizeof(int), 1, f);
-        fread(&e->id_graphe,  sizeof(int), 1, f);
+        fread(&e->id_article, sizeof(long), 1, f);
+        fread(&e->id_graphe,  sizeof(long), 1, f);
         HASH_ADD_INT(g->map_id, id_article, e);
     }
